### 2021-06-15 #3371 @whedon
[PRE REVIEW]: torchquad: Numerical Integration in Arbitrary Dimensions with PyTorch
**Submitting author:** @gomezzz (<a href="http://orcid.org/0000-0002-5631-8240">Pablo Gómez</a>)
**Repository:** <a href="https://github.com/esa/torchquad/" target ="_blank">https://github.com/esa/torchquad/</a>
**Version:** v0.2.1
