+++ paper/paper.bib
@@ -63,7 +63,8 @@
   volume={240},
   pages={120--137},
   year={2019},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/j.cpc.2019.02.015}
 }
 
 @article{Baydin2018autodiffinML,
@@ -96,7 +97,8 @@ author = {G.~Peter Lepage}
     year={2020},
     eprint={2009.05112},
     archivePrefix={arXiv},
-    primaryClass={physics.comp-ph}
+    primaryClass={physics.comp-ph},
+    doi={10.1016/j.jcp.2021.110386}
 }
 
 @incollection{PyTorch2019,
