### 2021-06-22 #3371 @gomezzz
@jbytecode Thanks for the list! After talking to my co-authors, from the list, given their expertise in numerical methods, machine learning and Python, we would suggest as potential reviewers

danehkar
felipefr
jakryd
jamiehadd
krystophny

Additionally, we would suggest the maintainers of the VegasFlow package which serves a similar purpose for the TensorFlow framework

scarrazza
scarlehoff

and the maintainers of PyTorch given their expertise with integrating with PyTorch and machine learning in general

apaszke
colesbury
gchanan
soumith

Took the liberty to provide a few more than necessary in case any of them are unavailable. :)
