### 2021-06-22 #3371 @felipefr
Yes, I wish to review. How can I do?

On Tue, 22 Jun 2021 at 14:34, Mehmet Hakan Satman ***@***.***>
wrote:

> @felipefr <https://github.com/felipefr> @jamiehadd
> <https://github.com/jamiehadd> Hi!
>
> Would you willing to assist in reviewing this for JOSS?
>
