### 2021-06-27 bea42a1
Update simpson.py
