### 2021-06-29 5a7de16
Update base_integrator.py
