+++ torchquad/integration/boole.py
@@ -11,13 +11,12 @@ logger = logging.getLogger(__name__)
 
 
 class Boole(BaseIntegrator):
-
     """Boole's rule in torch. See https://en.wikipedia.org/wiki/Newton%E2%80%93Cotes_formulas#Closed_Newton%E2%80%93Cotes_formulas ."""
 
     def __init__(self):
         super().__init__()
 
-    def integrate(self, fn, dim, N=None, integration_domain=None):
+    def integrate(self,fn,dim,N = None,integration_domain = None):
         """Integrates the passed function on the passed domain using Boole's rule.
 
         Args:
@@ -34,23 +33,18 @@ class Boole(BaseIntegrator):
         if N is None:
             N = 5 ** dim
 
-        self._integration_domain = setup_integration_domain(dim, integration_domain)
-        self._check_inputs(dim=dim, N=N, integration_domain=self._integration_domain)
-        N = self._adjust_N(dim=dim, N=N)
+        self._integration_domain = setup_integration_domain(dim,integration_domain)
+        self._check_inputs(dim = dim,N = N,integration_domain = self._integration_domain)
+        N = self._adjust_N(dim = dim,N = N)
 
         self._dim = dim
         self._fn = fn
 
-        logger.debug(
-            "Using Boole for integrating a fn with a total of "
-            + str(N)
-            + " points over "
-            + str(self._integration_domain)
-            + "."
-        )
+        logger.debug("Using Boole for integrating a fn with a total of " + str(N) + " points over " + str(
+            self._integration_domain) + ".")
 
         # Create grid and assemble evaluation points
-        self._grid = IntegrationGrid(N, self._integration_domain)
+        self._grid = IntegrationGrid(N,self._integration_domain)
 
         logger.debug("Evaluating integrand on the grid.")
         function_values = self._eval(self._grid.points)
@@ -65,51 +59,59 @@ class Boole(BaseIntegrator):
 
         # We collapse dimension by dimension
         for cur_dim in range(dim):
-            cur_dim_areas = (
-                self._grid.h[cur_dim]
-                / 22.5
-                * (
-                    7 * cur_dim_areas[..., 0:-4][..., ::4]
-                    + 32 * cur_dim_areas[..., 1:-3][..., ::4]
-                    + 12 * cur_dim_areas[..., 2:-2][..., ::4]
-                    + 32 * cur_dim_areas[..., 3:-1][..., ::4]
-                    + 7 * cur_dim_areas[..., 4:][..., ::4]
-                )
-            )
-            cur_dim_areas = torch.sum(cur_dim_areas, dim=dim - cur_dim - 1)
+            cur_dim_areas = (self._grid.h[cur_dim] / 22.5 * (
+                    7 * cur_dim_areas[...,0:-4][...,::4] + 32 * cur_dim_areas[...,1:-3][...,::4] + 12 * cur_dim_areas[
+            ...,2:-2][...,::4] + 32 * cur_dim_areas[...,3:-1][...,::4] + 7 * cur_dim_areas[...,4:][...,::4]))
+            cur_dim_areas = torch.sum(cur_dim_areas,dim = dim - cur_dim - 1)
         logger.info("Computed integral was " + str(cur_dim_areas) + ".")
 
         return cur_dim_areas
 
         @staticmethod
-        def _adjust_N(dim, N):
-        
+        def _adjust_N(dim,N):
+
         Args:
-            dim (int): Dimensionality of the integration domain.
-            N (int): Total number of sample points to use for the integration.
+        dim(int): Dimensionality
+        of
+        the
+        integration
+        domain.
+        N(int): Total
+        number
[…]
