### 2021-06-29 884bb2d
Update boole.py
