+++ torchquad/integration/boole.py
@@ -7,16 +7,16 @@ import torch
 import logging
 import warnings
 
-logger = logging.getLogger(__name__)
+logger = logging.getLogger( __name__ )
 
 
-class Boole(BaseIntegrator):
+class Boole( BaseIntegrator ) :
     """Boole's rule in torch. See https://en.wikipedia.org/wiki/Newton%E2%80%93Cotes_formulas#Closed_Newton%E2%80%93Cotes_formulas ."""
 
-    def __init__(self):
-        super().__init__()
+    def __init__( self ) :
+        super( ).__init__( )
 
-    def integrate(self,fn,dim,N = None,integration_domain = None):
+    def integrate( self ,fn ,dim ,N = None ,integration_domain = None ) :
         """Integrates the passed function on the passed domain using Boole's rule.
 
         Args:
@@ -30,53 +30,54 @@ class Boole(BaseIntegrator):
         """
 
         # If N is unspecified, set N to 5 points per dimension
-        if N is None:
+        if N is None :
             N = 5 ** dim
 
-        self._integration_domain = setup_integration_domain(dim,integration_domain)
-        self._check_inputs(dim = dim,N = N,integration_domain = self._integration_domain)
-        N = self._adjust_N(dim = dim,N = N)
+        self._integration_domain = setup_integration_domain( dim ,integration_domain )
+        self._check_inputs( dim = dim ,N = N ,integration_domain = self._integration_domain )
+        N = self._adjust_N( dim = dim ,N = N )
 
         self._dim = dim
         self._fn = fn
 
-        logger.debug("Using Boole for integrating a fn with a total of " + str(N) + " points over " + str(
-            self._integration_domain) + ".")
+        logger.debug( "Using Boole for integrating a fn with a total of " + str( N ) + " points over " + str(
+            self._integration_domain ) + "." )
 
         # Create grid and assemble evaluation points
-        self._grid = IntegrationGrid(N,self._integration_domain)
+        self._grid = IntegrationGrid( N ,self._integration_domain )
 
-        logger.debug("Evaluating integrand on the grid.")
-        function_values = self._eval(self._grid.points)
+        logger.debug( "Evaluating integrand on the grid." )
+        function_values = self._eval( self._grid.points )
 
         # Reshape the output to be [N,N,...] points instead of [dim*N] points
-        function_values = function_values.reshape([self._grid._N] * dim)
+        function_values = function_values.reshape( [ self._grid._N ] * dim )
 
-        logger.debug("Computing areas.")
+        logger.debug( "Computing areas." )
 
         # This will contain the Simpson's areas per dimension
         cur_dim_areas = function_values
 
         # We collapse dimension by dimension
-        for cur_dim in range(dim):
-            cur_dim_areas = (self._grid.h[cur_dim] / 22.5 * (
-                    7 * cur_dim_areas[...,0:-4][...,::4] + 32 * cur_dim_areas[...,1:-3][...,::4] + 12 * cur_dim_areas[
-            ...,2:-2][...,::4] + 32 * cur_dim_areas[...,3:-1][...,::4] + 7 * cur_dim_areas[...,4:][...,::4]))
-            cur_dim_areas = torch.sum(cur_dim_areas,dim = dim - cur_dim - 1)
-        logger.info("Computed integral was " + str(cur_dim_areas) + ".")
+        for cur_dim in range( dim ) :
+            cur_dim_areas = (self._grid.h[ cur_dim ] / 22.5 * (
[…]
