+++ .github/workflows/run_tests.yml
@@ -7,9 +7,11 @@ on:
   push:
     branches:
     - main
+    - develop
   pull_request:
     branches:
     - main
+    - develop
 jobs:
   build:
 
