### 2021-06-29 fa547d0
Update boole.py
