+++ torchquad/integration/boole.py
@@ -1,118 +1,46 @@
-from .base_integrator import BaseIntegrator
-from .integration_grid import IntegrationGrid
-from .utils import setup_integration_domain
+import sys
+
+sys.path.append("../")
 
 import torch
-
-import logging
 import warnings
 
-logger = logging.getLogger( __name__ )
+from integration.boole import Boole
+from utils.enable_cuda import enable_cuda
+from utils.set_precision import set_precision
 
 
-class Boole( BaseIntegrator ) :
-    """Boole's rule in torch. See https://en.wikipedia.org/wiki/Newton%E2%80%93Cotes_formulas#Closed_Newton%E2%80%93Cotes_formulas ."""
+def test_integrate():
+    """Tests the integrate function in integration.Boole.
+    Note: For now the 10-D test is diabled due to lack of GPU memory on some computers."""
+    enable_cuda()
+    set_precision("double")
 
-    def __init__( self ) :
-        super( ).__init__( )
+    # Needs to happen after precision / device settings to avoid having some tensors intialized on cpu and some on GPU
+    from tests.integration_test_utils import compute_test_errors
 
-    def integrate( self ,fn ,dim ,N = None ,integration_domain = None ) :
-        """Integrates the passed function on the passed domain using Boole's rule.
+    bl = Boole()
+    N = 401
 
-        Args:
-            fn (func): The function to integrate over.
-            dim (int): Dimensionality of the integration domain.
-            N (int, optional): Total number of sample points to use for the integration. N has to be such that N^(1/dim) - 1 % 4 == 0. Defaults to 5 points per dimension if None is given.
-            integration_domain (list, optional): Integration domain, e.g. [[-1,1],[0,1]]. Defaults to [-1,1]^dim.
+    errors = compute_test_errors(bl.integrate, {"N": N, "dim": 1})
+    print("Passed N =", N, "\n", errors)
+    for error in errors:
+        assert error < 1e-11
 
-        Returns:
-            float: integral value
-        """
+    # 3D Tests
+    N = 1076890  # N = 102.5 per dim (will change to 101 if all works)
+    with warnings.catch_warnings():
+        warnings.simplefilter("ignore")
+        errors = compute_test_errors(bl.integrate, {"N": N, "dim": 3}, dim = 3)
+    print("Passed N =", N, "\n", errors)
+    for error in errors[:3]:
+        assert error < 2e-13
+    for error in errors:
+        assert error < 5e-6
 
-        # If N is unspecified, set N to 5 points per dimension
-        if N is None :
-            N = 5 ** dim
-
-        self._integration_domain = setup_integration_domain( dim ,integration_domain )
-        self._check_inputs( dim = dim ,N = N ,integration_domain = self._integration_domain )
-        N = self._adjust_N( dim = dim ,N = N )
-
-        self._dim = dim
-        self._fn = fn
-
-        logger.debug( "Using Boole for integrating a fn with a total of " + str( N ) + " points over " + str(
-            self._integration_domain ) + "." )
-
-        # Create grid and assemble evaluation points
-        self._grid = IntegrationGrid( N ,self._integration_domain )
-
-        logger.debug( "Evaluating integrand on the grid." )
-        function_values = self._eval( self._grid.points )
-
-        # Reshape the output to be [N,N,...] points instead of [dim*N] points
-        function_values = function_values.reshape( [ self._grid._N ] * dim )
-
[…]
