+++ torchquad/tests/integration_test_functions.py
@@ -1,6 +1,6 @@
-import torch
-
 import logging
+
+import torch
 
 logger = logging.getLogger(__name__)
 
@@ -15,7 +15,7 @@ class IntegrationTestFunction:
     order = None
     f = None  # Function to evaluate
 
-    def __init__(self, expected_result, dim=1, domain=None):
+    def __init__(self, expected_result, dim = 1, domain = None):
         """Initializes domain and stores vars
 
         Args:
@@ -26,19 +26,13 @@ class IntegrationTestFunction:
         self.dim = dim
         self.expected_result = expected_result
         # Init domain to [-1,1]^dim if not passed
-        if domain == None:
-            self.domain = torch.tensor([[-1, 1]] * self.dim)
-        else:
-            self.domain = domain
+        self.domain = (
+            torch.tensor([[-1, 1]] * self.dim) if domain is None else domain
+        )
+
         logging.debug("Initialized Test function with ")
         logging.debug(
-            "dim="
-            + str(self.dim)
-            + "| domain="
-            + str(self.domain)
-            + "| expected_result="
-            + str(expected_result)
-        )
+            "dim=" + str(self.dim) + "| domain=" + str(self.domain) + "| expected_result=" + str(expected_result))
 
     def evaluate(self, integrator, integration_args):
         """Evaluates the passed integration functions with args
@@ -50,11 +44,11 @@ class IntegrationTestFunction:
         Returns:
             float: Integration result
         """
-        return integrator(fn=self.f, integration_domain=self.domain, **integration_args)
+        return integrator(fn = self.f, integration_domain = self.domain, **integration_args)
 
 
 class Polynomial(IntegrationTestFunction):
-    def __init__(self, expected_result=None, coeffs=[2], dim=1, domain=None):
+    def __init__(self, expected_result = None, coeffs = [ 2 ], dim = 1, domain = None):
         """N-dimensional , degree K poylnomial test functions
 
         Args:
@@ -70,22 +64,22 @@ class Polynomial(IntegrationTestFunction):
 
     def _poly(self, x):
         # compute x^k
-        exponentials = torch.zeros([x.shape[0], x.shape[1], self.order + 1])
+        exponentials = torch.zeros([ x.shape[ 0 ], x.shape[ 1 ], self.order + 1 ])
         for o in range(self.order + 1):
-            exponentials[:, :, o] = torch.pow(x, o)
+            exponentials[ :, :, o ] = torch.pow(x, o)
 
         # multiply by coefficients
         exponentials = torch.multiply(exponentials, self.coeffs)
 
         # Collapse dimensions
-        exponentials = torch.sum(exponentials, dim=2)
+        exponentials = torch.sum(exponentials, dim = 2)
 
         # sum all values for each dim
-        return torch.sum(exponentials, dim=1)
+        return torch.sum(exponentials, dim = 1)
 
 
 class Exponential(IntegrationTestFunction):
-    def __init__(self, expected_result=None, dim=1, domain=None):
+    def __init__(self, expected_result = None, dim = 1, domain = None):
         """Creates an n-dimensional exponential test function
 
         Args:
@@ -98,11 +92,11 @@ class Exponential(IntegrationTestFunction):
 
     def _exp(self, x):
         # compute e^x
-        return torch.sum(torch.exp(x), dim=1)
+        return torch.sum(torch.exp(x), dim = 1)
 
 
 class Sinusoid(IntegrationTestFunction):
-    def __init__(self, expected_result=None, dim=1, domain=None):
+    def __init__(self, expected_result = None, dim = 1, domain = None):
         """Creates an n-dimensional exponential test function
 
         Args:
[…]
