### 2021-06-29 #3371 @jbytecode
Dear @apaszke, @colesbury,  @gchanan, @soumith would one of you would be willing to assist in reviewing this submission for JOSS? I am sorry if I bothered you, I am just pinging you just because you are an expert in the research area of this submission.
