### 2021-06-29 29edac2
Code cleanup and fixed some issues
