### 2021-07-01 #3371 @jbytecode
Dear @scarlehoff , Thank you very much! The review process of JOSS is unique. Whenever we assign the reviewers, a new issue will be created. There will be twenty checkbox items will guide you. Whole the process is interactive between editors, reviewers, and the authors. Here is the reviewing guidelines: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html

Please, do not hesitate to consult me at all stages.
