### 2021-07-01 #3439 @jbytecode
Dear @felipefr and @scarlehoff,

This is the review issue for the submission. There are 20 checkbox items for each viewer. Please check the boxes during your review and make comments here. Since the review process is continuous rather than discrete, you can comment minor changes during the process so the authors can add/delete/correct changes simultaneously. You can also open new issues in the submitted repository and put links here.

Please do not hesitate to ask for help in any stage.

Thank you in advance for your contributions to JOSS!
