### 2021-07-05 esa/torchquad#89 @gomezzz
Initial feedback from JOSS review
> 1. There's a grammatical mistake here: https://github.com/esa/torchquad/blob/2824f1ddacdccadf3a880851f7ddf1bc1f878a36/torchquad/integration/base_integrator.py#L73external should be either "needs to match" or "needs to be match to". In either case, since integration_domain is referred by variable name, ndim would be better to be explictly mention as well

>   2. The usage example is clear and it works, but it should include the enable_cuda() call explicitly so that the user knows, just reading the readme, how to use the GPU-capabilities of the library (which is one of the main points)

>   3. From the readme and rtd, it seems to work only with CUDA/nvidia. Is that the case? It should be mentioned explicitly in the readme and docs. Note that pytorch offers also a rocm version so users might expect it to work with their AMD cards.

We have received initial feedback from [JOSS reviewers ](https://github.com/openjournals/joss-reviews/issues/3439) which we should address.
