### 2021-07-05 #3439 @gomezzz
Yes we can see it. :)

Thank you for the feedback! We will make those changes until the end of the week! I took the liberty to create an issue for it at https://github.com/esa/torchquad/issues/89
