### 2021-07-06 #3439 @scarlehoff
Everything seems to work as advertised. I don't have any important comments beyond the 3 that I added above.
The code is also quite clean and commented enough that it can be understood.

I haven't checked the performance box because while there's no direct claim to performance there are some indirect claims (and at the end of the day, that's what one would expect from a GPU code).
It would be nice to either provide a more complicated integral to exemplify this or to do some benchmark, not necessarily against other tools*. To be more specific, a comparison of torchquad running in 1-core, vs N-cores, Vs GPU would already be helpful to users wanting to know what kind of scaling are they supposed to see  in e.g., `VEGAS` and `Simpson`.

If the authors think such a comparison is beyond the scope (being the main goal just having a torch-based integrator) that's just as well. As we have found with VegasFlow, just the possibility of having a NN-compatible and GPU-compatible integrator already enables many studies and ideas that would not be easily possible otherwise. Performance can come later and I don't think it should be a showstopper (also, even without an explicit comparison, I would expect great performance gains for complicated integrands just by being able to use a GPU).

Anyway, as a summary: nice job ;)

*I'll add the "shameless plug" of Vegasflow with Vegas enhanced :P https://github.com/N3PDF/vegasflow/pull/64
