### 2021-07-07 8cb1a64
- Fixed some references to latest on rtd instead of main - Added mention of no AMD support atm - Added enable_cuda() to minimal example - Added more comments to minimal example - Small bugfix in minimal example - Fixed typo in base_integrator
