### 2021-07-07 7a54dc1
Apply suggestions from code review
Co-authored-by: Håvard Hem Toftevaag <70584234+htoftevaag@users.noreply.github.com>
