+++ docs/source/install.rst
@@ -16,7 +16,7 @@ Prerequisites
 We recommend using `conda <https://docs.conda.io/en/latest/>`_, especially if you want to utilize the GPU. 
 It will automatically set up CUDA and the cudatoolkit for you in that case.
 Note that *torchquad* also works on the CPU; however, it is optimized for GPU usage. 
-Currently torchquad only supports NVIDIA cards with CUDA. We are investigating future support AMD cards through [ROCm](https://pytorch.org/blog/pytorch-for-amd-rocm-platform-now-available-as-python-package/).
+Currently torchquad only supports NVIDIA cards with CUDA. We are investigating future support for AMD cards through [ROCm](https://pytorch.org/blog/pytorch-for-amd-rocm-platform-now-available-as-python-package/).
 
 For a detailed list of required packages, please refer to the `conda environment file <https://github.com/esa/torchquad/blob/main/environment.yml>`_.
 
