### 2021-07-15 #3439 @scarlehoff
> wave @scarlehoff, please update us on how your review is going (this is an automated reminder).

it is done. And the changes I asked for have been implemented I think https://github.com/esa/torchquad/pull/90
