### 2021-07-16 c9b87b3
*Need to mention that we autoformat every PR with black * Add that PRs should be created on develop not main branch due to our branching model * Add that only PRs which pass all tests will be considered
Only in README for now (also need docs)
