+++ requirements.txt
@@ -1,4 +1,4 @@
 matplotlib>=3.3.3
-torch==1.7.1
+torch==1.9.0
 scipy>=1.6.0
 tqdm>=4.56.0
