### 2021-07-16 69aa2b2
Added "Running Tests" build check
