+++ README.md
@@ -3,10 +3,12 @@
 *** Based on https://github.com/othneildrew/Best-README-Template
 -->
 
-[![Documentation Status](https://readthedocs.org/projects/torchquad/badge/?version=latest)](https://torchquad.readthedocs.io/en/latest/?badge=latest) &emsp; ![GitHub Workflow Status (branch)](https://img.shields.io/github/workflow/status/esa/torchquad/Running%20tests/develop) ![Conda](https://img.shields.io/conda/dn/conda-forge/torchquad) ![PyPI - Downloads](https://img.shields.io/pypi/dm/torchquad)
+[![Documentation Status](https://readthedocs.org/projects/torchquad/badge/?version=latest)](https://torchquad.readthedocs.io/en/latest/?badge=latest) ![GitHub Workflow Status (branch)](https://img.shields.io/github/workflow/status/esa/torchquad/Running%20tests/develop)  &emsp; 
+![Conda](https://img.shields.io/conda/dn/conda-forge/torchquad) ![PyPI - Downloads](https://img.shields.io/pypi/dm/torchquad)  &emsp; 
 ![GitHub issues](https://img.shields.io/github/issues/esa/torchquad) ![GitHub pull requests](https://img.shields.io/github/issues-pr/esa/torchquad) 
 
-![GitHub](https://img.shields.io/github/license/esa/torchquad) ![Conda (channel only)](https://img.shields.io/conda/vn/conda-forge/torchquad) ![PyPI](https://img.shields.io/pypi/v/torchquad) ![PyPI - Python Version](https://img.shields.io/pypi/pyversions/torchquad)
+![GitHub](https://img.shields.io/github/license/esa/torchquad)  &emsp; 
+![Conda (channel only)](https://img.shields.io/conda/vn/conda-forge/torchquad) ![PyPI](https://img.shields.io/pypi/v/torchquad) ![PyPI - Python Version](https://img.shields.io/pypi/pyversions/torchquad) &emsp; 
 ![GitHub last commit](https://img.shields.io/github/last-commit/esa/torchquad) ![GitHub contributors](https://img.shields.io/github/contributors/esa/torchquad)
 
 <!-- PROJECT LOGO -->
