### 2021-07-16 04f0767
Restructure shields
