### 2021-07-16 2667330
Try changing to square shields, remove whitespace
