### 2021-07-16 fd9377b
Square shields
