### 2021-07-16 e2f9b91
Added benchmark plots to readme
