+++ .github/ISSUE_TEMPLATE/feature_request.md
@@ -1,0 +1,15 @@
+---
+name: Feature request
+about: Suggest an idea for this project
+title: ''
+labels: ''
+assignees: ''
+
+---
+
+# Feature
+## Desired Behavior / Functionality
+
+## What Needs to Be Done
+
+## How Can It Be Tested
