+++ .github/PULL_REQUEST_TEMPLATE.md
@@ -1,0 +1,18 @@
+# Description
+
+Summary of changes
+
+* ... 
+* ...
+## Resolved Issues
+
+- [ ] fixes #Issue
+
+## How Has This Been Tested?
+
+- [ ] Test A
+- [ ] Test B
+
+## Related Pull Requests
+
+- #PR
