### 2021-07-16 04dc220
Create image shields for RtD
