### 2021-07-16 9335d01
- Adapted image path for main branch
