### 2021-07-16 a28a4e5
Apply suggestions from code review
Good suggestions, Pablo

Co-authored-by: Pablo Gómez <contact@pablo-gomez.net>
