### 2021-07-16 646d152
Apply suggestions from code review
Co-authored-by: Håvard Hem Toftevaag <70584234+htoftevaag@users.noreply.github.com>
