### 2021-07-16 16d9305
-fix whitespace in README - add to contributing.rst RTD
