### 2021-07-16 #3439 @gomezzz
> it is done. And the changes I asked for have been implemented I think esa/torchquad#90

Thank you for the review! Yes, you are right. We have implemented those changes and just today added some minor benchmark plots to the readme as well. Thanks! :)
