### 2021-07-16 890ce35
Apply suggestions from code review
Co-authored-by: Pablo Gómez <contact@pablo-gomez.net>
