+++ docs/source/contributing.rst
@@ -26,15 +26,15 @@ c. Setup the environment. This will create a conda environment called ``torchqua
       conda activate torchquad
 
 Once the installation is done, then you are ready to contribute. 
-Please note that PRs should be created from and into the ``develop`` branch. For each release the develop branch is moved into main.
+Please note that PRs should be created from and into the ``develop`` branch. For each release the develop branch is merged into main.
 
 3. Create your Feature Branch: ``git checkout -b feature/AmazingFeature``
 4. Commit your Changes: ``git commit -m 'Add some AmazingFeature'``
 5. Push to the Branch: ``git push origin feature/AmazingFeature``
-6. Open a Pull Request on the ``develop`` branch, *not* ``main`` (NB: For torchquad, we autoformat every PR with black. Our GitHub actions may create additional commits on your PR for that reason.)
+6. Open a Pull Request on the ``develop`` branch, *not* ``main`` (NB: We autoformat every PR with black. Our GitHub actions may create additional commits on your PR for that reason.)
 
 and we will have a look at your contribution as soon as we can. 
 
 Furthermore, please make sure that your PR passes all automated tests. Review will only happen after that.
-Only PRs created on the ``develop`` branch with all tests passed will be considered. The only exception to this rule is if you want to update the documentation in relation to the current release on ``conda`` / ``pip``. 
+Only PRs created on the ``develop`` branch with all tests passing will be considered. The only exception to this rule is if you want to update the documentation in relation to the current release on ``conda`` / ``pip``. 
 In that case you may ask to merge directly into ``main``.
