### 2021-07-23 8e015fe
- Removed some unused imports - Renamed a function to match naming conventions - Added gradient test (currently failing for VEGAS) - Added function to create a linspace without breaking gradients
