+++ torchquad/integration/utils.py
@@ -6,7 +6,34 @@ import logging
 logger = logging.getLogger(__name__)
 
 
-def setup_integration_domain(dim, integration_domain):
+def _linspace_with_grads(start, stop, N, requires_grad):
+    """Creates a equally spaced 1D grid while keeping gradients
+    in regard to inputs
+
+    Args:
+        start (torch.tensor): Start point (inclusive)
+        stop (torch.tensor): End point (inclusive)
+        N (torch.tensor): Number of points
+        requires_grad (bool): Indicates if output should be
+            recorded for backpropagation
+
+    Returns:
+        torch.tensor: Equally spaced 1D grid
+    """
+    if requires_grad:
+        # Create 0 to 1 spaced grid
+        grid = torch.linspace(0, 1, N)
+
+        # Scale to desired range , thus keeping gradients
+        grid *= stop - start
+        grid += start
+
+        return grid
+    else:
+        return torch.linspace(start, stop, N)
+
+
+def _setup_integration_domain(dim, integration_domain):
     """Sets up the integration domain if unspecified by the user.
 
     Args:
