### 2021-07-23 eb0276a
Test complex compability and fix test typos - Fixed difference in printout for tests - Fixed punctiation in docstrings in Test folder - Added complex functions to do tests on
