+++ torchquad/tests/trapezoid_test.py
@@ -19,17 +19,17 @@ def test_integrate():
     from tests.integration_test_utils import compute_test_errors
 
     tp = Trapezoid()
-    N = 100000
 
     # 1D Tests
+    N = 100000
     errors = compute_test_errors(tp.integrate, {"N": N, "dim": 1})
-    print("N =", N, "\n", errors)
+    print("Passed N =", N, "\n", "Errors: ", errors)
     for error in errors:
         assert error < 1e-5
 
     N = 2  # integration points, here 2 for order check (2 points should lead to almost 0 err for low order polynomials)
     errors = compute_test_errors(tp.integrate, {"N": N, "dim": 1})
-    print("N =", N, "\n", errors)
+    print("Passed N =", N, "\n", "Errors: ", errors)
     # all polynomials up to degree = 1 should be 0
     # if this breaks check if test functions in integration_test_utils changed.
     for error in errors[:2]:
@@ -38,16 +38,16 @@ def test_integrate():
     # 3D Tests
     N = 1000000
     errors = compute_test_errors(tp.integrate, {"N": N, "dim": 3}, dim=3)
-    print("N =", N, "\n", errors)
+    print("Passed N =", N, "\n", "Errors: ", errors)
     for error in errors[:2]:
         assert error < 1e-12
     for error in errors:
         assert error < 6e-3
 
-    # 10D Test
+    # 10D Tests
     N = 10000
     errors = compute_test_errors(tp.integrate, {"N": N, "dim": 10}, dim=10)
-    print("N =", N, "\n", errors)
+    print("Passed N =", N, "\n", "Errors: ", errors)
     for error in errors:
         assert error < 7000
 
