+++ torchquad/tests/vegas_test.py
@@ -25,7 +25,7 @@ def test_integrate():
     # 1D Tests
     N = 10000
     errors = compute_test_errors(vegas.integrate, {"N": N, "dim": 1, "seed": 0})
-    print("N =", N, "\n", errors)
+    print("Passed N =", N, "\n", "Errors: ", errors)
     for error in errors[:3]:
         assert error < 5e-3
 
@@ -38,16 +38,16 @@ def test_integrate():
     # 3D Tests
     N = 10000
     errors = compute_test_errors(vegas.integrate, {"N": N, "dim": 3, "seed": 0}, dim=3)
-    print("N =", N, "\n", errors)
+    print("Passed N =", N, "\n", "Errors: ", errors)
     for error in errors:
         assert error < 0.61
 
-    # 10D Test
+    # 10D Tests
     N = 10000
     errors = compute_test_errors(
         vegas.integrate, {"N": N, "dim": 10, "seed": 0}, dim=10
     )
-    print("N =", N, "\n", errors)
+    print("Passed N =", N, "\n", "Errors: ", errors)
     for error in errors:
         assert error < 12.5
 
