### 2021-07-23 626f7e5
Update run_tests.yml
Trying to fix training workflow not running on all PRs
