+++ .github/workflows/run_tests.yml
@@ -9,9 +9,7 @@ on:
     - main
     - develop
   pull_request:
-    branches:
-    - main
-    - develop
+  workflow_dispatch:
 jobs:
   build:
 
