### 2021-07-23 b8a519f
Update run_tests.yml
Also updating this workflow on develop branch
