+++ .github/workflows/run_tests.yml
@@ -7,9 +7,9 @@ on:
   push:
     branches:
     - main
+    - develop
   pull_request:
-    branches:
-    - main
+  workflow_dispatch:
 jobs:
   build:
 
