### 2021-07-30 e818ee3
Update docs in torchquad/integration/utils.py
