### 2021-07-30 245031d
- Changed test function name to avoid it being picked up by pytest - Added output which integrator is running
