+++ torchquad/tests/vegas_test.py
@@ -13,7 +13,7 @@ from utils.set_precision import set_precision
 
 
 def test_integrate():
-    """Tests the integrate function in integration.Trapezoid"""
+    """Tests the integrate function in integration.VEGAS."""
     enable_cuda()
     set_precision("double")
 
@@ -25,7 +25,7 @@ def test_integrate():
     # 1D Tests
     N = 10000
     errors = compute_test_errors(vegas.integrate, {"N": N, "dim": 1, "seed": 0})
-    print("N =", N, "\n", errors)
+    print("Passed N =", N, "\n", "Errors: ", errors)
     for error in errors[:3]:
         assert error < 5e-3
 
@@ -38,22 +38,22 @@ def test_integrate():
     # 3D Tests
     N = 10000
     errors = compute_test_errors(vegas.integrate, {"N": N, "dim": 3, "seed": 0}, dim=3)
-    print("N =", N, "\n", errors)
+    print("Passed N =", N, "\n", "Errors: ", errors)
     for error in errors:
         assert error < 0.61
 
-    # 10D Test
+    # 10D Tests
     N = 10000
     errors = compute_test_errors(
         vegas.integrate, {"N": N, "dim": 10, "seed": 0}, dim=10
     )
-    print("N =", N, "\n", errors)
+    print("Passed N =", N, "\n", "Errors: ", errors)
     for error in errors:
         assert error < 12.5
 
 
 if __name__ == "__main__":
-    # used to run this test individually
+    # Used to run this test individually
     start = timeit.default_timer()
     test_integrate()
     # cProfile.run("test_integrate()")
