+++ torchquad/tests/simpson_test.py
@@ -11,7 +11,7 @@ from utils.set_precision import set_precision
 
 
 def test_integrate():
-    """Tests the integrate function in integration.Simpson"""
+    """Tests the integrate function in integration.Simpson."""
     enable_cuda()
     set_precision("double")
 
@@ -19,16 +19,18 @@ def test_integrate():
     from tests.integration_test_utils import compute_test_errors
 
     simp = Simpson()
+
+    # 1D Tests
     N = 100001
 
     errors = compute_test_errors(simp.integrate, {"N": N, "dim": 1})
-    print("Passed N =", N, "\n", errors)
+    print("Passed N =", N, "\n", "Errors: ", errors)
     for error in errors:
         assert error < 1e-7
 
     N = 3  # integration points, here 3 for order check (3 points should lead to almost 0 err for low order polynomials)
     errors = compute_test_errors(simp.integrate, {"N": N, "dim": 1})
-    print("Passed N =", N, "\n", errors)
+    print("Passed N =", N, "\n", "Errors: ", errors)
     # all polynomials up to degree = 3 should be 0
     # if this breaks check if test functions in integration_test_utils changed.
     for error in errors[:3]:
@@ -39,13 +41,13 @@ def test_integrate():
     with warnings.catch_warnings():
         warnings.simplefilter("ignore")
         errors = compute_test_errors(simp.integrate, {"N": N, "dim": 3}, dim=3)
-    print("Passed N =", N, "\n", errors)
+    print("Passed N =", N, "\n", "Errors: ", errors)
     for error in errors[:3]:
         assert error < 1e-12
     for error in errors:
         assert error < 5e-6
 
-    # 10D Test
+    # 10D Tests
     N = 3 ** 10
     errors = compute_test_errors(simp.integrate, {"N": N, "dim": 10}, dim=10)
     print("N =", N, "\n", errors)
