+++ torchquad/tests/boole_test.py
@@ -20,10 +20,11 @@ def test_integrate():
     from tests.integration_test_utils import compute_test_errors
 
     bl = Boole()
+    # 1D Tests
     N = 401
 
     errors = compute_test_errors(bl.integrate, {"N": N, "dim": 1})
-    print("Passed N =", N, "\n", errors)
+    print("Passed N =", N, "\n", "Errors: ", errors)
     for error in errors:
         assert error < 1e-11
 
@@ -32,17 +33,17 @@ def test_integrate():
     with warnings.catch_warnings():
         warnings.simplefilter("ignore")
         errors = compute_test_errors(bl.integrate, {"N": N, "dim": 3}, dim=3)
-    print("Passed N =", N, "\n", errors)
+    print("Passed N =", N, "\n", "Errors: ", errors)
     for error in errors[:3]:
         assert error < 2e-13
     for error in errors:
         assert error < 5e-6
 
-    # 10D Test
+    # 10D Tests
     # Have been disabled for now because it is too GPU-heavy
     # N = 5 ** 10
     # errors = compute_test_errors(bl.integrate, {"N": N, "dim": 10}, dim=10)
-    # print("N =", N, "\n", errors)
+    # print("Passed N =", N, "\n", "Errors: ", errors)
     # for error in errors:
     # assert error < 5e-9
 

[…]
