### 2021-07-30 5d19f19
Added “scaffolding” for complex numbers in set_precision and enable_cuda
