### 2021-07-30 6f40ea5
Added examples how to extract gradients to the documentation.
