### 2021-07-30 d3c0da7
Moved complex_test to normal tests - Added use_complex to the individual test functions - Added is_complex to the class IntegrationTestFunction - Checked that all tests run
