+++ torchquad/tests/vegas_test.py
@@ -24,8 +24,10 @@ def test_integrate():
 
     # 1D Tests
     N = 10000
-    errors = compute_test_errors(vegas.integrate, {"N": N, "dim": 1, "seed": 0})
-    print("Passed N =", N, "\n", "Errors: ", errors)
+    errors = compute_test_errors(
+        vegas.integrate, {"N": N, "dim": 1, "seed": 0}, use_complex=False
+    )
+    print("1D VEGAS Test: Passed N =", N, "\n", "Errors: ", errors)
     for error in errors[:3]:
         assert error < 5e-3
 
@@ -37,17 +39,19 @@ def test_integrate():
 
     # 3D Tests
     N = 10000
-    errors = compute_test_errors(vegas.integrate, {"N": N, "dim": 3, "seed": 0}, dim=3)
-    print("Passed N =", N, "\n", "Errors: ", errors)
+    errors = compute_test_errors(
+        vegas.integrate, {"N": N, "dim": 3, "seed": 0}, dim=3, use_complex=False
+    )
+    print("3D VEGAS Test: Passed N =", N, "\n", "Errors: ", errors)
     for error in errors:
         assert error < 0.61
 
     # 10D Tests
     N = 10000
     errors = compute_test_errors(
-        vegas.integrate, {"N": N, "dim": 10, "seed": 0}, dim=10
+        vegas.integrate, {"N": N, "dim": 10, "seed": 0}, dim=10, use_complex=False
     )
-    print("Passed N =", N, "\n", "Errors: ", errors)
+    print("10D VEGAS Test: Passed N =", N, "\n", "Errors: ", errors)
     for error in errors:
         assert error < 12.5
 
