+++ torchquad/tests/boole_test.py
@@ -23,8 +23,8 @@ def test_integrate():
     # 1D Tests
     N = 401
 
-    errors = compute_test_errors(bl.integrate, {"N": N, "dim": 1})
-    print("Passed N =", N, "\n", "Errors: ", errors)
+    errors = compute_test_errors(bl.integrate, {"N": N, "dim": 1}, use_complex=True)
+    print("1D Boole Test: Passed N =", N, "\n", "Errors: ", errors)
     for error in errors:
         assert error < 1e-11
 
@@ -32,8 +32,10 @@ def test_integrate():
     N = 1076890  # N = 102.5 per dim (will change to 101 if all works)
     with warnings.catch_warnings():
         warnings.simplefilter("ignore")
-        errors = compute_test_errors(bl.integrate, {"N": N, "dim": 3}, dim=3)
-    print("Passed N =", N, "\n", "Errors: ", errors)
+        errors = compute_test_errors(
+            bl.integrate, {"N": N, "dim": 3}, dim=3, use_complex=True
+        )
+    print("3D Boole Test: Passed N =", N, "\n", "Errors: ", errors)
     for error in errors[:3]:
         assert error < 2e-13
     for error in errors:
@@ -42,8 +44,8 @@ def test_integrate():
     # 10D Tests
     # Have been disabled for now because it is too GPU-heavy
     # N = 5 ** 10
-    # errors = compute_test_errors(bl.integrate, {"N": N, "dim": 10}, dim=10)
-    # print("Passed N =", N, "\n", "Errors: ", errors)
+    # errors = compute_test_errors(bl.integrate, {"N": N, "dim": 10}, dim=10, use_complex=True)
+    # print("10D Boole Test: Passed N =", N, "\n", "Errors: ", errors)
     # for error in errors:
     # assert error < 5e-9
 

[…]
