### 2021-07-30 6548e54
Added detach() to error assertion
