### 2021-07-30 6a17716
Various optimizations
