### 2021-07-30 #3439 @jbytecode
> > wave @scarlehoff, please update us on how your review is going (this is an automated reminder).
>
> it is done. And the changes I asked for have been implemented I think [esa/torchquad#90](https://github.com/esa/torchquad/pull/90)

Dear @scarlehoff

Thank you for your great effort. Additionally, please just check the checkbox if you don't think the software claims any performance issues so all of your reviewing tasks will be completed (20/20).

Thank you in advance.
