### 2021-07-30 c8bf927
Update docs/source/tutorial.rst
Co-authored-by: Håvard Hem Toftevaag <70584234+htoftevaag@users.noreply.github.com>
