### 2021-07-30 0cba9ad
Update docs/source/tutorial.rst
Co-authored-by: Håvard Hem Toftevaag <70584234+htoftevaag@users.noreply.github.com>
