### 2021-07-30 8eea4c2
Update docs/source/tutorial.rst
Co-authored-by: Håvard Hem Toftevaag <70584234+htoftevaag@users.noreply.github.com>
