+++ docs/source/tutorial.rst
@@ -406,7 +406,7 @@ Reasonably, one might prefer Monte Carlo integration methods for a 5-D
 problem. We might add this comparison to the tutorial in the future.
 
 Computing gradients with respect to the integration domain
---------------------------------------------------------
+----------------------------------------------------------
 
 *torchquad* allows fully automatic differentiation. In this tutorial, we will show how to extract the gradients with respect to the integration domain.
 We selected Trapezoid and MonteCarlo methods to showcase that getting the gradients it is possible for both deterministic and stochastic methods.
