### 2021-08-06 85f9745
Update cudatoolkit version - RtD fails if not
