+++ environment.yml
@@ -3,7 +3,7 @@ channels:
 - conda-forge
 - pytorch
 dependencies:
-- cudatoolkit>=11.0.221
+- cudatoolkit>=11.1
 - matplotlib>=3.3.3
 - pytest>=6.2.1
 - python>=3.8
