### 2021-08-06 766fdcd
- Switch to loguru - Small fixes - Add some feedback on loading torchquad - Cleaned up some unused imports
