+++ torchquad/tests/monte_carlo_test.py
@@ -7,10 +7,11 @@ import torch
 from integration.monte_carlo import MonteCarlo
 from utils.enable_cuda import enable_cuda
 from utils.set_precision import set_precision
-
+from utils.set_log_level import set_log_level
 
 def test_integrate():
     """Tests the integrate function in integration.MonteCarlo."""
+    set_log_level("INFO")
     enable_cuda()
     set_precision("double")
 
