+++ torchquad/integration/vegas_map.py
@@ -1,9 +1,5 @@
 import torch
-from copy import deepcopy
 
-import logging
-
-logger = logging.getLogger(__name__)
 
 
 class VEGASMap:
