+++ torchquad/integration/monte_carlo.py
@@ -1,11 +1,8 @@
 from .base_integrator import BaseIntegrator
-from .integration_grid import IntegrationGrid
 from .utils import _setup_integration_domain
 import torch
 
-import logging
-
-logger = logging.getLogger(__name__)
+from loguru import logger
 
 
 class MonteCarlo(BaseIntegrator):
