### 2021-08-06 7611822
Added JOSS badge to RtD
