### 2021-08-06 08e8bbe
WORKFLOW: Format Python code with black
