+++ torchquad/tests/monte_carlo_test.py
@@ -9,6 +9,7 @@ from utils.enable_cuda import enable_cuda
 from utils.set_precision import set_precision
 from utils.set_log_level import set_log_level
 
+
 def test_integrate():
     """Tests the integrate function in integration.MonteCarlo."""
     set_log_level("INFO")
