+++ environment.yml
@@ -3,7 +3,7 @@ channels:
 - conda-forge
 - pytorch
 dependencies:
-- cudatoolkit>=11.0.221
+#- cudatoolkit>=11.0.221
 - loguru>=0.5.3
 - matplotlib>=3.3.3
 - pytest>=6.2.1
