+++ .github/workflows/run_tests.yml
@@ -27,7 +27,7 @@ jobs:
         activate-environment: ""
     - name: Install dependencies
       run: |
-        conda env create --file environment.yml --name base
+        conda env create --file environment.yml
     - name: Lint with flake8
       run: |
         # stop the build if there are Python syntax errors or undefined names
