+++ .github/workflows/run_tests.yml
@@ -37,6 +37,7 @@ jobs:
     #     flake8 . --count --exit-zero --max-complexity=10 --max-line-length=127 --statistics
     - name: Test with pytest
       run: |
+        conda activate torchquad
         cd torchquad/tests
         conda install pytest
-        $CONDA/bin/pytest
+        pytest
