### 2021-08-06 5ec11cc
Fixed bug in gradient flow
