+++ torchquad/tests/gradient_test.py
@@ -28,9 +28,9 @@ def test_gradients():
     enable_cuda()
     set_precision("double")
 
-    N = 199997
+    N = 99997
     result_tolerence = 1e-2
-    gradient_tolerance = 5e-2
+    gradient_tolerance = 2e-2
     torch.manual_seed(0)  # we have to seed torch to get reproducible results
 
     # Define integrators
