### 2021-08-06 ce1135f
Taking a few steps back and trying to setup loguru on github actions properly
