+++ .github/workflows/run_tests.yml
@@ -21,24 +21,19 @@ jobs:
       uses: actions/setup-python@v2
       with:
         python-version: 3.9
-    - uses: conda-incubator/setup-miniconda@v2
-      with:
-        auto-activate-base: true
-        activate-environment: ""
     - name: Install dependencies
       run: |
-        conda env create --file environment.yml
+        $CONDA/bin/conda env update --file environment.yml --name base
     - name: Lint with flake8
       run: |
+        pip install flake8
         # stop the build if there are Python syntax errors or undefined names
-        conda activate torchquad
-        conda install flake8
         flake8 . --count --select=E9,F63,F7,F82 --show-source --statistics
     #     # exit-zero treats all errors as warnings. The GitHub editor is 127 chars wide
     #     flake8 . --count --exit-zero --max-complexity=10 --max-line-length=127 --statistics
     - name: Test with pytest
       run: |
-        conda activate torchquad
         cd torchquad/tests
+        pip install loguru
         conda install pytest
-        pytest
+        $CONDA/bin/pytest
