+++ torchquad/integration/simpson.py
@@ -1,11 +1,12 @@
+import torch
+from loguru import logger
+import warnings
+
 from .base_integrator import BaseIntegrator
 from .integration_grid import IntegrationGrid
 from .utils import _setup_integration_domain
 
-import torch
 
-from loguru import logger
-import warnings
 
 
 class Simpson(BaseIntegrator):
