+++ torchquad/utils/enable_cuda.py
@@ -1,7 +1,7 @@
 import torch
 import os
-
 from loguru import logger
+
 from .set_precision import set_precision
 
 
