+++ torchquad/integration/trapezoid.py
@@ -1,10 +1,9 @@
+import torch
+from loguru import logger
+
 from .base_integrator import BaseIntegrator
 from .integration_grid import IntegrationGrid
 from .utils import _setup_integration_domain
-
-import torch
-
-from loguru import logger
 
 
 class Trapezoid(BaseIntegrator):
