+++ .github/workflows/run_tests.yml
@@ -21,9 +21,6 @@ jobs:
       uses: actions/setup-python@v2
       with:
         python-version: 3.9
-    - name: Install dependencies
-      run: |
-        $CONDA/bin/conda env update --file environment.yml --name base
     - name: Lint with flake8
       run: |
         pip install flake8
@@ -31,9 +28,11 @@ jobs:
         flake8 . --count --select=E9,F63,F7,F82 --show-source --statistics
     #     # exit-zero treats all errors as warnings. The GitHub editor is 127 chars wide
     #     flake8 . --count --exit-zero --max-complexity=10 --max-line-length=127 --statistics
+    - name: Install dependencies
+      run: |
+        $CONDA/bin/conda env update --file environment.yml --name base
     - name: Test with pytest
       run: |
         cd torchquad/tests
-        pip install loguru
         conda install pytest
         $CONDA/bin/pytest
