### 2021-08-06 d318828
WORKFLOW: Format Python code with black
