+++ torchquad/utils/get_logger.py
@@ -1,0 +1,4 @@
+from loguru import logger
+
+def get_logger():
+    return logger
