### 2021-08-06 155c912
WORKFLOW: Format Python code with black
