+++ torchquad/utils/get_logger.py
@@ -1,4 +1,5 @@
 from loguru import logger
 
+
 def get_logger():
     return logger
