+++ torchquad/utils/enable_cuda.py
@@ -1,10 +1,8 @@
 import torch
 import os
+from loguru import logger
 
 from .set_precision import set_precision
-from .get_logger import get_logger
-
-logger = get_logger()
 
 
 def enable_cuda(device=0, data_type="float"):
@@ -24,6 +22,6 @@ def enable_cuda(device=0, data_type="float"):
         logger.info("Active CUDA Device: GPU" + str(torch.cuda.current_device()))
         set_precision(data_type)
     else:
-        logger.warn(
+        logger.warning(
             "Error enabling CUDA. cuda.is_available() returned False. CPU will be used."
         )
