### 2021-08-09 #3439 @felipefr
I think _torchquad_ is already a very mature library, with equally reasonably complete documentation. I definitively agree that it matches the levels of software innovation and design required by JOSS. My only suggestion is that the reported _issues_ are still too limited, https://github.com/esa/torchquad/issues reporting more of them would attract more contributors.
