### 2021-08-09 #3439 @jbytecode
Dear @gomezzz

- [ ] In paper, some function and package names are written in different font (using texttt) but some of them are written in regular font (pip and Python in paper, torchquad in summary, and many...). Please adopt a single rule for function, package, and software names. It is convenient to use `texttt` for all of them.

- [ ] Please re-check the whole paper by means of both language and format. When it is finished, let whedon generate the pdf and check references here.

- [ ] make a tagged release and archive (e.g. with [Zenodo](https://zenodo.org/)) as described [here](<https://joss.readthedocs.io/en/latest/submitting.html#the-review-process>), and report the version number and archive DOI in this thread. Please verify that the archive deposit has the correct metadata (title and author list), or edit these if that is not the case.

Thank you in advance.
