+++ paper/paper.md
@@ -1,55 +1,26 @@
----
-title: 'torchquad: Numerical Integration in Arbitrary Dimensions with PyTorch'
-tags:
-  - Python
-  - n-dimensional
-  - numerical integration
-  - GPU
-  - automatic differentiation
-  - PyTorch
-  - high-performance computing
-  - machine learning
-authors:
-  - name: Pablo Gómez^[corresponding author]
-    orcid: 0000-0002-5631-8240
-    affiliation: 1
-  - name: Håvard Hem Toftevaag
-    orcid: 0000-0003-4692-5722
-    affiliation: 1
-  - name: Gabriele Meoni
-    orcid: 0000-0001-9311-6392
-    affiliation: 1
-affiliations:
- - name: Advanced Concepts Team, European Space Agency, Noordwijk, The Netherlands
-   index: 1
-date: 15 June 2021
-bibliography: paper.bib
-
----
-
 # Summary
 
-\texttt{torchquad} is a Python module for $n$-dimensional numerical integration optimized for graphics processing units (GPUs).
-Various deterministic and stochastic integration methods, such as Newton\textendash Cotes formulas and Monte Carlo integration methods like VEGAS Enhanced [@VegasEnhanced-paper], are available for computationally efficient integration for arbitrary dimensionality $n_{\mathrm{d}}$.
-As it is implemented using PyTorch [@PyTorch2019], one of the most popular machine learning frameworks, \texttt{torchquad} provides fully automatic differentiation throughout the integration, which is essential for many machine learning applications.
+\texttt{torchquad} is a \texttt{Python} module for $n$-dimensional numerical integration optimized for graphics processing units (GPUs).
+Various deterministic and stochastic integration methods, such as \texttt{Newton\textendash Cotes} formulas and \texttt{Monte Carlo} integration methods like \texttt{VEGAS Enhanced} [@VegasEnhanced-paper], are available for computationally efficient integration for arbitrary dimensionality $n_{\mathrm{d}}$.
+As it is implemented using \texttt{PyTorch} [@PyTorch2019], one of the most popular machine learning frameworks, \texttt{torchquad} provides fully automatic differentiation throughout the integration, which is essential for many machine learning applications.
 
 # Statement of Need
 
 Multidimensional integration is needed in many fields, such as physics (ranging from particle physics [@ParticlePhysics-Paper] to astrophysics [@izzo2021geodesy]), applied finance [@AppliedFinance-Paper], medical statistics [@MedicalStatistics-Paper], and machine learning [@VEGASinMachineLearning-Paper]. 
-Most of the conventional Python packages for multidimensional integration, such as \texttt{quadpy} [@quadpy] and \texttt{nquad} [@scipy], only target and are optimized for central processing units (CPUs). 
+Most of the conventional \texttt{Python} packages for multidimensional integration, such as \texttt{quadpy} [@quadpy] and \texttt{nquad} [@scipy], only target and are optimized for central processing units (CPUs). 
 However, as many numerical integration methods are embarrassingly parallel, GPUs can offer superior computational performance in their computation. 
 Furthermore, numerical integration methods typically suffer from the so-called \textit{curse of dimensionality} [@ZMCintegral]. 
 This phenomenon refers to the fact that the computational complexity of the integration grows exponentially with the number of dimensions [@CurseOfDim-Book]. Reducing the error of the integration value requires increasing the number of function evaluation points $N$ exponentially, which significantly increases the runtime of the computation, especially for higher dimensions.
 Previous work has demonstrated that this problem can be mitigated by leveraging the \textit{single instruction, multiple data} parallelization of GPUs [@ZMCintegral].
 
[…]
