+++ docs/source/tutorial.rst
@@ -24,10 +24,10 @@ abritrary dimensionality.
 | Trapezoid    | Creates a linear interpolant between two |br|   | Equal      |
 | rule         | neighbouring points                             |            |
 +--------------+-------------------------------------------------+------------+
-| Simpson’s    | Creates a quadratic interpolant between |br|    | Equal      |
+| Simpson's    | Creates a quadratic interpolant between |br|    | Equal      |
 | rule         | three neighbouring point                        |            |
 +--------------+-------------------------------------------------+------------+
-| Boole’s      | Creates a more complex interpolant between |br| | Equal      |
+| Boole's      | Creates a more complex interpolant between |br| | Equal      |
 | rule         | five neighbouring points                        |            |
 +--------------+-------------------------------------------------+------------+
 | Monte Carlo  | Randomly chooses points at which the |br|       | Random     |
@@ -49,10 +49,11 @@ Outline
 This notebook is a guide for new users to *torchquad* and is structured in
 the following way:
 
--  Necessary imports for this tutorial
--  Example integration in one dimension (1-D) with complex numbers
--  Example integration in ten dimensions (10-D)
--  Some accuracy / runtime comparisons with scipy
+1.  Necessary imports for this tutorial
+2.  Example integration in one dimension (1-D) with complex numbers
+3.  Example integration in ten dimensions (10-D)
+4.  Some accuracy / runtime comparisons with scipy
+5.  Example showing how gradients can be obtained w.r.t. the integration domain
 
 Feel free to test the code on your own computer as we go along.
 
@@ -95,53 +96,58 @@ Now let’s get started! First, the general imports:
 
 
 
-One-dimensional integration
-----------------------------
+One-dimensional integration with complex numbers
+------------------------------------------------
 
 To make it easier to understand the methods used in this notebook, we will start with an
 example in one dimension. If you are new to these methods or simply want a clearer picture, 
 feel free to check out Patrick Walls’ 
 `nice Python introduction <https://github.com/patrickwalls/mathematical-python/>`__ 
 to the `Trapezoid rule <https://www.math.ubc.ca/~pwalls/math-python/integration/trapezoid-rule/>`__
-and `Simpson’s rule <https://www.math.ubc.ca/~pwalls/math-python/integration/simpsons-rule/>`__
+and `Simpson's rule <https://www.math.ubc.ca/~pwalls/math-python/integration/simpsons-rule/>`__
 in one dimension.
 Similarly, `Tirthajyoti Sarkar <https://github.com/tirthajyoti>`__ has made a nice visual explanation of 
 `Monte Carlo integration in Python 
 <https://towardsdatascience.com/monte-carlo-integration-in-python-a71a209d277e>`__.
 
-The *torchquad* package also works with complex numbers. To showcase this 
-
-Let ``f(x)`` be the function :math:`f(x) = e^{x} \cdot x^{2}`. Over the domain 
+The *torchquad* package also works with complex numbers. To showcase this, in the following example we 
+will let ``f(x)`` be the complex function :math:`f(x) = e^{3ix} \cdot x^{2}`. Over the domain 
 :math:`[0,2]`, the integral of ``f(x)`` is :math:`\int_{0}^{2} f(x) dx = 
-\int_{0}^{2} e^x \cdot x^2 dx = 2(e^{2} - 1) = 12.7781121978613004544...`
+\int_{0}^{2} e^{3ix} \cdot x^{2} dx = -\frac{2}{27} i ( 1 + (17 + 6i) e^{6i} )`, or, alternatively, 
+:math:`0.0748857930... - i1.4073621035...` 
[…]
