### 2021-08-13 #3439 @gomezzz
> I think _torchquad_ is already a very mature library, with equally reasonably complete documentation. I definitively agree that it matches the levels of software innovation and design required by JOSS. My only suggestion is that the reported _issues_ are still too limited, https://github.com/esa/torchquad/issues reporting more of them would attract more contributors.

Thank you for your review! We are very happy to receive your positive feedback! :)

You are right about the _issues_. To alleviate, we will soon transition tickets from our JIRA to GitHub _issues_ and try to mark some issues as good first issues for interested contributors.

Thanks!
