### 2021-08-13 #3439 @gomezzz
Dear @jbytecode ,

we have applied the requested formatting ([here](https://github.com/esa/torchquad/pull/114)) , rechecked the paper and uploaded the repository (torchquad 0.2.1) to Zenodo.

On Zenodo, it can be found [here](https://zenodo.org/record/5195695#.YRZHhYgzYuU). The DOI is 10.5281/zenodo.5195695 .

Thanks!
