### 2021-08-18 #3439 @gomezzz
Hi @jbytecode,
sorry for not getting back to you earlier. Unfortunately, none of us are an admin of the ESA GitHub organization under which torchquad is published and thus we are unable to link the repository directly on Zenodo as we require organization level access for this. We have contacted the admins but they are currently on vacation. If there is an alternative way to archive the repository please let us know. Otherwise, we will have to wait until they get back to us.

Sorry for the inconvenience!
