+++ setup.py
@@ -9,7 +9,7 @@ from setuptools import setup, find_packages
 
 setup(
     name="torchquad",
-    version="0.2.2",
+    version="0.2.3",
     description="Package providing torch-based numerical integration methods.",
     long_description=open("README.md").read(),
     long_description_content_type="text/markdown",
