### 2021-08-20 95b1746
Several minor fixes - Fixed install.rst link failure - Updated typos in tutorial.rst - Updated and fixed output in tutorial.rst - Fixed utils.py docstring and comment - Fixed enable_cuda.py print
