### 2021-08-20 e9337ac
Delete utils.py
This will be added in another PR
