### 2021-08-20 9561ad6
Delete enable_cuda.py
This will be added in another PR
