### 2021-08-20 1e2c24c
Apply suggestions from code review
Added Pablo's comments

Co-authored-by: Pablo Gómez <contact@pablo-gomez.net>
