### 2021-08-20 3f4ca74
Suggested changes to tutorial.rst
