+++ docs/source/tutorial.rst
@@ -82,6 +82,7 @@ Now let’s get started! First, the general imports:
     
     from torchquad import enable_cuda # Necessary to enable GPU support
     from torchquad import Trapezoid, Simpson, Boole, MonteCarlo, VEGAS # The available integrators
+    from torchquad.utils.set_precision import set_precision
     import torchquad
 
 .. code:: ipython3
@@ -153,12 +154,13 @@ Let’s plot the function briefly.
 .. image:: torchquad_tutorial_figure.png
 
 
-Let’s define the integration domain now and initialize the integrator - let’s start with the trapezoid rule.
+Let’s define the integration domain, set the precision to double, and initialize the integrator - let’s start with the trapezoid rule.
 
 .. code:: ipython3
 
     # Integration domain is always a list of lists to allow arbitrary dimensionality.
     integration_domain = [[0, 2]]
+    set_precision('double')
     tp = Trapezoid()  # Initialize a trapezoid solver
 
 Now we are all set to compute the integral. Let’s try it with just 101 sample points for now.
@@ -171,9 +173,9 @@ Now we are all set to compute the integral. Let’s try it with just 101 sample 
 
 .. parsed-literal::
 
-    **Output**: Results: (0.07512567937374115-1.407015323638916j)
-            Abs. Error: 4.21665376e-04
-            Rel. Error: 2.99190753e-04
+    **Output**: Results: (0.0751256138749595-1.4070152346569569j)
+            Abs. Error: 4.21725301e-04
+            Rel. Error: 2.99233274e-04
     
 
 This is quite close already, as 1-D integrals are comparatively easy.
@@ -188,9 +190,9 @@ Let’s see what type of value we get for different integrators.
 
 .. parsed-literal::
 
-    **Output:** Results: (0.0748857706785202-1.4073622226715088j)
-            Abs. Error: 1.21286661e-07
-            Rel. Error: 8.60583995e-08
+    **Output:** Results: (0.0748856817342774-1.407362129881349j)
+            Abs. Error: 1.11758709e-07
+            Rel. Error: 7.92978838e-08
     
 .. code:: ipython3
 
@@ -201,9 +203,9 @@ Let’s see what type of value we get for different integrators.
 
 .. parsed-literal::
 
-    **Output:** Results: (0.0748857855796814-1.4073619842529297j)
-            Abs. Error: 1.19441893e-07
-            Rel. Error: 8.47494519e-08
+    **Output:** Results: (0.07488579656243394-1.407362063802251j)
+            Abs. Error: 0.00000000e+00
+            Rel. Error: 0.00000000e+00
 
 
 .. code:: ipython3
@@ -215,36 +217,23 @@ Let’s see what type of value we get for different integrators.
 
 .. parsed-literal::
 
-    **Output:** (0.050353918224573135-1.494397759437561j)
-            Abs. Error: 9.04268697e-02
-            Rel. Error: 6.41619712e-02
+    **Output:** Results: (0.1345164951606459-1.1738163554904042j)
+            Abs. Error: 2.41038278e-01
+            Rel. Error: 1.71027616e-01
     
 
-.. code:: ipython3
-
-    vegas = VEGAS()
-    result = vegas.integrate(f,dim=1,N=101,integration_domain=integration_domain)
-    print_error(result,solution)
-
-
-.. parsed-literal::
-
-    **Output:** RuntimeError: result type ComplexFloat can't be cast to the desired output type Float
-    
 
 Notably, Simpson's and Boole's methods are already really good here with only 101 points. 
-(You can try setting the function ``f(x) = torch.exp(x) * torch.pow(x,2)`` instead 
-to see Simpson's and Boole's methods find the perfect solution).
 Monte Carlo methods do not perform so well; they are more suited to higher-dimensional integrals. 
[…]
