### 2021-08-20 eeea508
Update tutorial.rst
Fixed "output"
