### 2021-08-31 #3439 @gomezzz
@jbytecode We have gotten access and created a tagged release v0.2.4 [here](https://github.com/esa/torchquad/releases/tag/v0.2.4) and linked Zenodo release [here](https://zenodo.org/record/5344938). The DOI for Zenodo is 10.5281/zenodo.5344938 . This should now be the correct format, right?

Thanks!

PS. Had to create a new release as Zenodo only picks up on new releases.
