### 2021-08-31 #3439 @jbytecode
@gomezzz the zenodo entry includes `Pablo Gómez; Håvard Hem Toftevaag; Gabriele Meoni; Albern S.` as authors; however, the paper has only three authors. The metadata should match in both side.
