### 2021-08-31 #3439 @gomezzz
@jbytecode I have updated the authors on Zenodo to match. (Also added ORCIDs on Zenodo)
