### 2021-09-01 f6d0552
Added paper link to README.md
