### 2021-09-01 38e1043
- Fixed warnings in docs - Added separate environment for readthedocs to fix failing builds
