+++ rtd_environment.yml
@@ -1,0 +1,15 @@
+name: rtd_torchquad
+channels:
+- conda-forge
+- pytorch
+dependencies:
+- ipython
+- loguru>=0.5.3
+- matplotlib>=3.3.3
+- pytest>=6.2.1
+- python>=3.8
+- pytorch>=1.9
+- scipy>=1.6.0
+- sphinx>=3.4.3
+- sphinx_rtd_theme>=0.5.1
+- tqdm>=4.56.0
