### 2021-09-01 cf22802
- Added separate environment for readthedocs to fix failing builds - Fixed warnings in docs
